test fixes and created tests for api and remote mediator

+++ app/src/main/java/com/snehil/falconix/api/model/Payload.kt
@@ -16,8 +16,8 @@ data class Payload (
   @SerializedName("nationality"      ) var nationality    : String?           = null,
   @SerializedName("manufacturer"     ) var manufacturer   : String?           = null,
   @SerializedName("payload_type"     ) var payloadType    : String?           = null,
-  @SerializedName("payload_mass_kg"  ) var payloadMassKg  : Int?              = null,
-  @SerializedName("payload_mass_lbs" ) var payloadMassLbs : Int?              = null,
+  @SerializedName("payload_mass_kg"  ) var payloadMassKg  : Double?              = null,
+  @SerializedName("payload_mass_lbs" ) var payloadMassLbs : Double?              = null,
   @SerializedName("orbit"            ) var orbit          : String?           = null,
   @Embedded
   @SerializedName("orbit_params"     ) var orbitParams    : OrbitParams?      = OrbitParams()
